requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
jinja2==3.1.4
requests==2.32.3
gtts==2.5.3
openai-whisper==20240930
ffmpeg-python==0.2.0